HTTP study
This is really almost all review stuff for me. I zeroed in on the stuff
I wasn't familiar with, but client server model and all of the nuts and
bolts of this were very much drilled into me in my classes for web
design. I am going to take a deep dive into some of the stuff that is
new to me, however - specifically the restful stuff.